Creating validation, error responses and lombok constructors

+++ src/main/java/io/github/pedromartinsl/libraryapi/service/AutorService.java
@@ -6,19 +6,23 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import io.github.pedromartinsl.libraryapi.exceptions.OperacaoNegadaException;
 import io.github.pedromartinsl.libraryapi.model.Autor;
 import io.github.pedromartinsl.libraryapi.repository.AutorRepository;
+import io.github.pedromartinsl.libraryapi.repository.LivroRepository;
+import io.github.pedromartinsl.libraryapi.validor.AutorValidator;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
-
-    public AutorService(AutorRepository repository) {
-        this.repository = repository;
-    }
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +30,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar é necessário que o autor já esteja necessário na base");
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -34,6 +39,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNegadaException("Não é permitido excluir um autor possui que livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -47,5 +55,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
